macOS (Beta)/requirements.txt
flask>=3.0.0
waitress>=3.0.0
rumps>=0.4.0
Pillow>=10.0.0
